refac(rust): make the lib more modularized based on thier tasks
