+++ apps/app/src-tauri/src/tauri_commands/mod.rs
@@ -1,0 +1,1 @@
+pub mod file_system_stuff;
